airflow-pipeline/requirements.txt
apache-airflow
boto3
tensorflow
pandas
numpy
joblib